backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.6.4
python-multipart==0.0.9

langchain==0.1.16
langchain-groq==0.1.3
